+++ src/game/board/cell/product/powershell/Powershell.view.fxml
@@ -1,0 +1,14 @@
+package Product.Powershell;
+
+public class Powershell extends Product {
+
+    public Powershell() {
+
+    }
+
+    @override
+    public void init_dice(){
+        //TODO fill the custom dice
+    }
+
+}
